Backend

+++ blog/views.py
@@ -3,23 +3,45 @@ from . forms import *
 from . models import *
 
 
+# def add_blog(request):
+#     if request.user.is_authenticated:
+#              if request.user.is_doctor:
+#                   if request.method =='POST':
+#                       form=BlogPostForm(request.POST,request.FILES)
+#                       if form.is_valid():
+#                              blog = form.save(commit=False)
+#                              blog.author = request.user
+#                              blog.save()
+#                              print("Add blog sucessfully")
+#                              return render(request,'dashboard_doctor.html')
+#                   else:
+#                      print('not working')
+#                      form=BlogPostForm()
+#     else:
+#          return redirect('login')
+#     return render(request,'add_blog.html',{'form':form})
 def add_blog(request):
-    if request.user.is_authenticated:
-             if request.user.is_doctor:
-                  if request.method =='POST':
-                      form=BlogPostForm(request.POST,request.FILES)
-                      if form.is_valid():
-                             blog = form.save(commit=False)
-                             blog.author = request.user
-                             blog.save()
-                             print("Add blog sucessfully")
-                             return render(request,'dashboard_doctor.html')
-                  else:
-                     print('not working')
-                     form=BlogPostForm()
+    if not request.user.is_authenticated:
+        return redirect('login')
+
+    if not request.user.is_doctor:
+        # Handle case where user is not a doctor, maybe show an error or redirect
+        return render(request, 'error.html', {'message': 'You are not authorized to add a blog.'})
+
+    if request.method == 'POST':
+        form = BlogPostForm(request.POST, request.FILES)
+        if form.is_valid():
+            blog = form.save(commit=False)
+            blog.author = request.user
+            blog.save()
+            print("Add blog successfully")
+            return redirect('dashboard')  # Assuming 'dashboard_doctor' is the name of the URL
+        else:
+            print('Form is invalid')
     else:
-         return redirect('login')
-    return render(request,'add_blog.html',{'form':form})
+        form = BlogPostForm()
+
+    return render(request, 'add_blog.html', {'form': form})
 
 def doctor_display_blog(request):
      blog=BlogPost.objects.all()
@@ -28,6 +50,13 @@ def doctor_display_blog(request):
 
 
 def update_blog(request,id):
+    if not request.user.is_authenticated:
+        return redirect('login')
+
+    if not request.user.is_doctor:
+        # Handle case where user is not a doctor, maybe show an error or redirect
+        return render(request, 'error.html', {'message': 'You are not authorized to add a blog.'})
+    
     blog = get_object_or_404(BlogPost, id=id)
     if request.method == 'POST':
         form = BlogPostForm(request.POST, request.FILES, instance=blog)
@@ -40,6 +69,12 @@ def update_blog(request,id):
     return render(request, 'update_blog.html', {'form': form, 'blog': blog})
 
 def delete_blog(request,id):
-        blog=get_object_or_404(BlogPost,id=id)
-        blog.delete()
-        return redirect('dashboard')
+    if not request.user.is_authenticated:
+        return redirect('login')
+
+    if not request.user.is_doctor:
+        # Handle case where user is not a doctor, maybe show an error or redirect
+        return render(request, 'error.html', {'message': 'You are not authorized to delete a blog.'})
+    blog=get_object_or_404(BlogPost,id=id)
+    blog.delete()
+    return redirect('dashboard')
